refactor: float 타입으로 변경 및 코드 정리

+++ Source/Core/Rendering/FDeviceResource.cpp
@@ -215,7 +215,7 @@ void FDevice::InitResource()
 		TArray<uint32> indices;
 		int slices = 16;
 		int stacks = 16;
-		int32 radius = 1.f;  // FIXME: why not float?
+		float radius = 1.f;
 		[[maybe_unused]] float height = 1.f;
 
 		UGeometryGenerator::CreateSphere(radius, slices, stacks, vertices, indices);
